backend/requirements.txt
django-environ==0.11.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
whitenoise>=6.7,<7

# API auth (JWT)
djangorestframework-simplejwt==5.3.1

# Filtering
django-filter==25.1

# Image handling (required by Django ImageField)
Pillow>=10,<11


django-storages
google-cloud-storage